Move DatabaseHolder classes from platform to internal package
Platform packages should only contain expect/actual declarations.
DatabaseHolder is a platform-specific implementation detail but not
an expect/actual, so it belongs in the internal package.
